+++ tests/test_tmy.py
@@ -23,8 +23,10 @@ def test_rooftop_pv_power():
 def test_fail_rooftop_pv_power():
     lats, longs = load_test_locations_coordinates()
 
-    res = tmy.rooftop_pv_power(latitude=lats[0], longitude=longs[0], array_type="wrong")
+    res = tmy.rooftop_pv_power(
+        latitude=lats[0], longitude=longs[0], array_type="wrong", capacity=3
+    )
 
     assert res.success is False
     assert res.code == 400
-    assert res.exception == "The specified condition was not met for 'Array Type'."
+    assert res.exception == "The specified condition was not met for 'array_type'."